Support ASPA payload in RTR. (#250)
This PR adds support for protocol version 2 of RTR. This introduces a new
PDU type for ASPA payload.

This PR also fixes a bug in the RTR server code that included router key
PDUs in the updates even if protocol version 0 was negotiated. (The client
code accepts all PDU types in all versions.)

The PR also changes the methods and functions in the rtr::pdu module to use
the type Asn instead of u32 for ASNs.