Various improvements and changes to support ASPA in the RTR server. (#252)
This PR includes a number of improvements and changes, some breaking, to
support ASPA in the RTR server.

Most importantly, it introduces a new dedicated type for references to any
type of RTR payload, PayloadRef, rather than using &Payload, so a reference
can be generated from a reference to a distinct payload type. The server
traits now use this new type to refer to payload.

This is a breaking change.