Return expected version in the RTR server error PDU on mismatch. (#280)
This PR changes the version reported by the RTR server in the header of the
RTR error PDU when a request has an invalid version. It returns the maximum
supported version if a version hasn’t been negotiated yet. It returns the
negotiated version if the client uses a different version after negotiation.

---------

Co-authored-by: Luuk Hendriks <github@luukhendriks.eu>