Correctly encode a CRL with empty revokedCertificates. (#295)
This PR fixes the encoding of CRLs by leaving out the revokedCertificates
element entirely if it is empty as required by section 5.1.2.6 of RFC 5280.

There is no test whether it is actually left out since I can’t quite think
of a way to do that. I have, however, manually verified the encoding and
there are now tests for both cases that check that we ourselves accept the
result.