Add limits to reading XML inputs (#319)
This PR adds limits to XML parsing to avoid memory exhaustion by ill-crafted
XML input. It adds “trip counter” to the reader that feeds data into the XML
parser which limits the size of an opening tag to 1,000,000 bytes and the
content of certain XML elements to 100,000,000 bytes. This, together with
the XML parser itself not keeping data buffered protects against exploding
compressed data.