IP Resources to prefix: fix overflow for /0 (#323)
Previously it would make a prefix, add that to the start, and check whether
the end is still smaller than the starting address. This can lead to
situations where e.g. 0.0.0.0 - 255.255.255.255 would overflow, as it would
try to go up to 256.0.0.0, which the u32 cannot contain.

These changes add a check for the last address to fix that.