Implement signing time attribute requirements from RFC 9589. (#340)
This PR implements the changes to signed objects from RFC 6488 introduced by
RFC 9589. Specifically, it makes the Signing Time signed attribute mandatory
and the Binary Signing Time attribute illegal. This results in some changes
to the APIs around signed objects, making this a breaking change.

I have tested the changes against a current data set of the public RPKI and
all signed objects fulfilled the new requirements, so these changes can be
applied to non-strict validation as well.

This PR may, however, have consequences for Krill, in particular when
dealing with old data. Careful testing is needed before upgrading Krill to
use this version of rpki-rs.

This lack of conformance with RFC 9589 was reported by Zhechao Lin, Zizhi
Shang, Jiahao Cao, Yangyang Wang, and Mingwei Xu of the Institute for
Network Sciences and Cyberspace, Tsinghua University.