Check correct critical flag in certificate extensions. (#339)
This PR changes the code parsing the certificate extension of resource
certificates to check that the critical flag for each is set to the value
required by RFC 5280 and 6487. It does not check the flag in certificate
signing requests or identity certificates to avoid potential surprise
interoperability issues in Krill.

I checked this change against the current data in the public RPKI and it did
not affect any object, so applying these stricter checks even in non-strict
mode should be fine.

This issue was reported by Zizhi Shang, Zhechao Lin, Jiahao Cao, Yangyang
Wang, and Mingwei Xu of the Institute for Network Sciences and Cyberspace
(INSC), Tsinghua University.