Fix Key Usage parsing. (#337)
This PR fixes parsing of Key Usage extension of resource certificates.
Previously, parsing only checked that the required bits were set but it
didn’t check that all other bits were zero.

I checked the change against the current version of the public RPKI and it
does not affect any current objects, so it should be safe to apply in
non-strict mode as well.

This issue was reported by Zizhi Shang, Zhechao Lin, Jiahao Cao, Yangyang
Wang, and Mingwei Xu of the Institute for Network Sciences and Cyberspace
(INSC), Tsinghua University.